Implemented a Credential System
Admin can see and modify every user and all tasks, while normal users are only able to modify themselves, but can see all the other users and their todos.

Also extracted a method to check whether an email is already taken or not in the UserService.

In TodoService extracted a method to check the credentials before modifying a TODO.