Implemented the profile picture operations
NOTE: The unidirectional connection is broken, so they are not linked together, you have to first delete the specific picture for the user, then you will be able to delete a user.

Introduced Lombok to reduce the amount of boilerplate code there is to be written.

In the application.properties I hid the database username and password -> now they are stored and accessed as environmental variables.

I fixed the user handling, some credential checks were buggy.

Introduced new exceptions as well.