Updated Tests and Fully implemented the Logger
In the previous Commit I mistakenly included everything, not just a readme.

I implemented and updated the unit and integration tests for the userService and userController classes.

For testing I used an H2 in-memory db.
Introduced a respective application-properties file for testing.

Extracted any error message into ErrorConstants, that was used more than once.

Refactored the EmailService class as well.